Add files via upload
Sentiment Analysis on Moroccan Arabic Dialect
Case study: Analysing Covid19 vaccination opinions collected from News web 
predict the sentiment of a user comment, whether it is a positive sentiment or negative. If so, predict a 1. If not, predict a 0.
The training set is not too large and not too small: it consists of 1920 comments, alongside a test set that contains 240 comments.
This dataset has been built using comments and opinions of users from news websites and Social Media about covid19 vaccination gathered by the class labeled by multiple annotators, the final labels assigned to each in the dataset were chosen by the majority vote on all the labels generated by the annotators.
There are only two classes of sentiment in this dataset which are Negative (0) and Positive(1).